Continue GUI implementation
Create lobby editor GUI with map loading and saving.

+++ src/main/java/me/mprey/util/WorldUtil.java
@@ -2,10 +2,8 @@ package me.mprey.util;
 
 import me.mprey.Annihilation;
 import me.mprey.map.MapManager;
-import org.bukkit.Difficulty;
-import org.bukkit.Location;
-import org.bukkit.World;
-import org.bukkit.WorldCreator;
+import org.bukkit.*;
+import org.bukkit.block.Block;
 import org.bukkit.generator.BlockPopulator;
 import org.bukkit.generator.ChunkGenerator;
 
@@ -49,6 +47,70 @@ public class WorldUtil {
         deleteWorld(new File(Annihilation.getInstance().getServer().getWorldContainer().getAbsolutePath(), worldName));
     }
 
+    public static World createEmptyWorld(String worldName) {
+        WorldCreator worldCreator = new WorldCreator(worldName);
+        worldCreator.environment(World.Environment.NORMAL);
+        worldCreator.generateStructures(false);
+        worldCreator.generator(new ChunkGenerator() {
+            @Override
+            public byte[] generate(World world, Random random, int x, int z) {
+                return super.generate(world, random, x, z);
+            }
+
+            @Override
+            public short[][] generateExtBlockSections(World world, Random random, int x, int z, BiomeGrid biomes) {
+                return super.generateExtBlockSections(world, random, x, z, biomes);
+            }
+
+            @Override
+            public byte[][] generateBlockSections(World world, Random random, int x, int z, BiomeGrid biomes) {
+                return super.generateBlockSections(world, random, x, z, biomes);
+            }
+
+            @Override
+            public ChunkData generateChunkData(World world, Random random, int x, int z, BiomeGrid biome) {
+                return super.generateChunkData(world, random, x, z, biome);
+            }
+
+            @Override
+            public boolean canSpawn(World world, int x, int z) {
+                return true;
+            }
+
+            @Override
+            public List<BlockPopulator> getDefaultPopulators(World world) {
+                return Arrays.asList(new BlockPopulator[0]);
+            }
+
+            @Override
+            public Location getFixedSpawnLocation(World world, Random random) {
+                return new Location(world, 0.0D, 64.0D, 0.0D);
+            }
+        });
+
+        World world = worldCreator.createWorld();
+        world.setDifficulty(Difficulty.NORMAL);
+        world.setSpawnFlags(true, true);
+        world.setPVP(!worldName.equals("lobby"));
+        world.setStorm(false);
+        world.setThundering(false);
+        world.setWeatherDuration(Integer.MAX_VALUE);
+        world.setAutoSave(false);
+        world.setKeepSpawnInMemory(false);
+        world.setTicksPerAnimalSpawns(1);
+        world.setTicksPerMonsterSpawns(1);
+
+        world.setGameRuleValue("doMobSpawning", "false");
+        world.setGameRuleValue("mobGriefing", "false");
+        world.setGameRuleValue("doFireTick", "false");
+        world.setGameRuleValue("showDeathMessages", "false");
+
+        Block b = world.getBlockAt(0, 20, 0);
+        b.setType(Material.STONE);
+
+        return world;
+    }
+
     public static boolean loadWorld(String worldName) {
         boolean isLobby = worldName.equalsIgnoreCase("lobby");
 
@@ -56,13 +118,28 @@ public class WorldUtil {
         creator.generateStructures(false);
         creator.generator(new ChunkGenerator() {
             @Override
-            public boolean canSpawn(World world, int x, int z) {
-                return true;
+            public byte[] generate(World world, Random random, int x, int z) {
+                return super.generate(world, random, x, z);
+            }
+
+            @Override
+            public short[][] generateExtBlockSections(World world, Random random, int x, int z, BiomeGrid biomes) {
+                return super.generateExtBlockSections(world, random, x, z, biomes);
+            }
+
+            @Override
+            public byte[][] generateBlockSections(World world, Random random, int x, int z, BiomeGrid biomes) {
+                return super.generateBlockSections(world, random, x, z, biomes);
             }
 
             @Override
             public ChunkData generateChunkData(World world, Random random, int x, int z, BiomeGrid biome) {
                 return super.generateChunkData(world, random, x, z, biome);
+            }
+
+            @Override
+            public boolean canSpawn(World world, int x, int z) {
+                return true;
             }
 
             @Override
@@ -106,7 +183,9 @@ public class WorldUtil {
                 if (f.isDirectory()) {
                     deleteWorld(f);
                 } else {
-                    f.delete();
+                    if (!f.getName().contains(".yml")) {
+                        f.delete();
+                    }
                 }
             }
         }
